+++ scoop/ColorPalette.cpp
@@ -20,6 +20,7 @@
  *
  */
 #include "ColorPalette.h"
+#include "ColorBrewerPalettes.h"
 #include <iostream>
 
 namespace scoop
@@ -39,52 +40,67 @@ namespace scoop
     return ( unsigned int ) _colors.size( );
   }
 
-  const char* cbBuGn[][9] =
-  {{ "#2ca25f", "#e5f5f9" },
-   { "#2ca25f", "#99d8c9", "#e5f5f9" },
-   { "#238b45", "#66c2a4", "#b2e2e2", "#edf8fb" },
-   { "#006d2c", "#2ca25f", "#66c2a4", "#b2e2e2", "#edf8fb" },
-   { "#006d2c", "#2ca25f", "#66c2a4", "#99d8c9", "#ccece6", "#edf8fb" },
-   { "#005824", "#238b45", "#41ae76", "#66c2a4", "#99d8c9", "#ccece6",
-     "#edf8fb" },
-   { "#005824", "#238b45", "#41ae76", "#66c2a4", "#99d8c9", "#ccece6",
-     "#e5f5f9", "#f7fcfd" },
-   { "#00441b", "#006d2c", "#238b45", "#41ae76", "#66c2a4", "#99d8c9",
-     "#ccece6", "#e5f5f9", "#f7fcfd" }};
-
-  const char* cbBuPu[][9] =
-  {{ "#8856a7", "#e0ecf4"},
-   { "#8856a7", "#9ebcda", "#e0ecf4"},
-   { "#88419d", "#8c96c6", "#b3cde3", "#edf8fb"},
-   { "#810f7c", "#8856a7", "#8c96c6", "#b3cde3", "#edf8fb"},
-   { "#810f7c", "#8856a7", "#8c96c6", "#9ebcda", "#bfd3e6", "#edf8fb"},
-   { "#6e016b", "#88419d", "#8c6bb1", "#8c96c6", "#9ebcda", "#bfd3e6",
-     "#edf8fb"},
-   { "#6e016b", "#88419d", "#8c6bb1", "#8c96c6", "#9ebcda", "#bfd3e6",
-     "#e0ecf4", "#f7fcfd"},
-   { "#4d004b", "#810f7c", "#88419d", "#8c6bb1", "#8c96c6", "#9ebcda",
-     "#bfd3e6", "#e0ecf4", "#f7fcfd"}};
-
   ColorPalette ColorPalette::colorBrewerSequential(
     ColorBrewerSequential cbPalette,
-    unsigned int size )
+    unsigned int size,
+    bool reverse )
   {
-    // char (*p)[][7][9];
-    if ( size > 9 || size < 2 )
+    if ( size > 9 || size < 3 )
       throw std::runtime_error( "Palette size invalid" );
 
     ColorPalette palette;
-    auto stringifiedPalette = cbBuGn;
-    if ( cbPalette == ColorBrewerSequential::BuGn )
-      stringifiedPalette = cbBuGn;
-    else if ( cbPalette == ColorBrewerSequential::BuPu )
-      stringifiedPalette = cbBuPu;
-    else
-      throw std::runtime_error( "Unknown palette" );
+    auto stringifiedPalette = cbSequential[ cbPalette ];
     auto& colors = palette.colors( );
     colors.reserve( size );
     for ( unsigned int i = 0; i < size; ++i )
-      colors.push_back( Color( stringifiedPalette[size-2][i] ));
+      colors.push_back( Color( stringifiedPalette[size-3][i] ));
+
+    if ( reverse )
+      std::reverse( colors.begin( ), colors.end( ));
+
     return palette;
   }
+
+  ColorPalette ColorPalette::colorBrewerDiverging(
+    ColorBrewerDiverging cbPalette,
+    unsigned int size,
+    bool reverse )
+  {
+    if ( size > 11 || size < 3 )
+      throw std::runtime_error( "Palette size invalid" );
+
+    ColorPalette palette;
+    auto stringifiedPalette = cbDiverging[ cbPalette ];
+    auto& colors = palette.colors( );
+    colors.reserve( size );
+    for ( unsigned int i = 0; i < size; ++i )
+      colors.push_back( Color( stringifiedPalette[size-3][i] ));
+
+    if ( reverse )
+      std::reverse( colors.begin( ), colors.end( ));
+
+    return palette;
+  }
+
+  ColorPalette ColorPalette::colorBrewerQualitative(
+    ColorBrewerQualitative cbPalette,
+    unsigned int size,
+    bool reverse )
+  {
+    if ( size > cbQualitativeSizes[cbPalette] || size < 3 )
+      throw std::runtime_error( "Palette size invalid" );
+
+    ColorPalette palette;
+    auto stringifiedPalette = cbQualitative[cbPalette];
+    auto& colors = palette.colors( );
+    colors.reserve( size );
+    for ( unsigned int i = 0; i < size; ++i )
+      colors.push_back( Color( stringifiedPalette[size-3][i] ));
+
+    if ( reverse )
+      std::reverse( colors.begin( ), colors.end( ));
+
+    return palette;
+  }
+
 }