+++ Changelog.md
@@ -2,3 +2,4 @@
 
 ## git master
 * [!1] Fixed building
+* [!2] Added ColorBrewer palettes (sequential, diverging and qualitative)